% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ads.R
\docType{data}
\name{ads_status_text}
\alias{ads_status_text}
\title{ADS job status codes}
\format{
An object of class \code{character} of length 5.
}
\usage{
ads_status_text
}
\description{
ADS job status codes
}
\keyword{internal}
