% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brightspaceR-package.R
\docType{package}
\name{brightspaceR-package}
\alias{brightspaceR}
\alias{brightspaceR-package}
\title{brightspaceR: Access D2L Brightspace Data Sets via the BDS API}
\description{
Connect to the D2L Brightspace Data Sets (BDS) API via OAuth2, download all available datasets as tidy data frames with proper types, join them using convenience functions that know the foreign key relationships, and analyse student engagement, performance, and retention with composable analytics functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pcstrategyandopsco.github.io/brightspaceR/}
  \item \url{https://github.com/pcstrategyandopsco/brightspaceR}
  \item Report bugs at \url{https://github.com/pcstrategyandopsco/brightspaceR/issues}
}

}
\author{
\strong{Maintainer}: Peeyoosh Chandra \email{peeyoosh.chandra@example.com}

}
\keyword{internal}
