% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bs_api_version}
\alias{bs_api_version}
\title{Get or set the Brightspace API version}
\usage{
bs_api_version(version = NULL)
}
\arguments{
\item{version}{If provided, sets the API version. If \code{NULL}, returns the
current version.}
}
\value{
Character string of the API version.
}
\description{
Get or set the Brightspace API version
}
\examples{
bs_api_version()
bs_api_version("1.49")
}
