% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{bs_auth_interactive}
\alias{bs_auth_interactive}
\title{Interactive browser-based OAuth2 flow}
\usage{
bs_auth_interactive(
  client_id,
  client_secret,
  instance_url,
  redirect_uri,
  scope
)
}
\description{
Builds the authorization URL, opens the browser, prompts for the redirect
URL, and exchanges the code for a token.
}
\keyword{internal}
