% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{bs_clean_names}
\alias{bs_clean_names}
\title{Convert column names from PascalCase to snake_case}
\usage{
bs_clean_names(df)
}
\arguments{
\item{df}{A data frame.}
}
\value{
A data frame with snake_case column names.
}
\description{
Convert column names from PascalCase to snake_case
}
\examples{
df <- data.frame(UserId = 1, FirstName = "A")
bs_clean_names(df)
}
