% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{bs_coerce_types}
\alias{bs_coerce_types}
\title{Coerce character columns to appropriate types}
\usage{
bs_coerce_types(df)
}
\arguments{
\item{df}{A tibble with all-character columns.}
}
\value{
A tibble with coerced types.
}
\description{
For unknown datasets, attempts to convert character columns to numeric,
logical, or datetime types.
}
\keyword{internal}
