% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{bs_deauth}
\alias{bs_deauth}
\title{Clear Brightspace authentication}
\usage{
bs_deauth(clear_cache = TRUE)
}
\arguments{
\item{clear_cache}{If \code{TRUE} (default), also removes the cached token from
disk.}
}
\value{
Invisibly returns \code{TRUE}.
}
\description{
Removes cached credentials from the current session and optionally from
disk.
}
\examples{
\dontrun{
bs_deauth()
}
}
