% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{bs_download}
\alias{bs_download}
\title{Download a file from a URL}
\usage{
bs_download(url, dest_path, download_size = NULL)
}
\arguments{
\item{url}{Full URL to download.}

\item{dest_path}{Destination file path.}
}
\value{
The destination path (invisibly).
}
\description{
Download a file from a URL
}
\keyword{internal}
