% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{bs_error_body}
\alias{bs_error_body}
\title{Extract error message from Brightspace API response body}
\usage{
bs_error_body(resp)
}
\arguments{
\item{resp}{An httr2 response.}
}
\value{
Character string with error details.
}
\description{
Extract error message from Brightspace API response body
}
\keyword{internal}
