% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{bs_get}
\alias{bs_get}
\title{Perform a GET request and return parsed JSON}
\usage{
bs_get(path, query = list())
}
\arguments{
\item{path}{API path.}

\item{query}{Named list of query parameters.}
}
\value{
Parsed JSON response as a list.
}
\description{
Perform a GET request and return parsed JSON
}
\keyword{internal}
