% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bs_get_timezone}
\alias{bs_get_timezone}
\title{Get the current Brightspace analytics timezone}
\usage{
bs_get_timezone()
}
\value{
Character string of the timezone.
}
\description{
Returns the timezone set by \code{\link[=bs_set_timezone]{bs_set_timezone()}}, defaulting to \code{"UTC"}.
}
\examples{
bs_get_timezone()
}
