% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{bs_has_token}
\alias{bs_has_token}
\title{Check if authenticated with Brightspace}
\usage{
bs_has_token()
}
\value{
Logical; \code{TRUE} if a token is available.
}
\description{
Check if authenticated with Brightspace
}
\examples{
bs_has_token()
}
