% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{bs_instance_url}
\alias{bs_instance_url}
\title{Get the Brightspace instance URL}
\usage{
bs_instance_url()
}
\value{
Character string of the instance URL.
}
\description{
Get the Brightspace instance URL
}
\keyword{internal}
