% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{bs_is_transient}
\alias{bs_is_transient}
\title{Determine if an HTTP error is transient}
\usage{
bs_is_transient(resp)
}
\arguments{
\item{resp}{An httr2 response.}
}
\value{
Logical.
}
\description{
Determine if an HTTP error is transient
}
\keyword{internal}
