% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{bs_join_content_progress}
\alias{bs_join_content_progress}
\title{Join content objects with user progress}
\usage{
bs_join_content_progress(content_objects, content_progress)
}
\arguments{
\item{content_objects}{A tibble from the Content Objects dataset.}

\item{content_progress}{A tibble from the Content User Progress dataset.}
}
\value{
A joined tibble.
}
\description{
Left joins a content objects tibble with a content user progress tibble
on \code{content_object_id} and \code{org_unit_id}.
}
\examples{
\donttest{
if (bs_has_token()) {
content <- bs_get_dataset("Content Objects")
progress <- bs_get_dataset("Content User Progress")
bs_join_content_progress(content, progress)
}
}
}
