% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{bs_join_enrollments_grades}
\alias{bs_join_enrollments_grades}
\title{Join enrollments with grade results}
\usage{
bs_join_enrollments_grades(enrollments, grade_results)
}
\arguments{
\item{enrollments}{A tibble from the User Enrollments dataset.}

\item{grade_results}{A tibble from the Grade Results dataset.}
}
\value{
A joined tibble.
}
\description{
Left joins an enrollments tibble with a grade results tibble on
\code{org_unit_id} and \code{user_id}.
}
\examples{
\donttest{
if (bs_has_token()) {
enrollments <- bs_get_dataset("User Enrollments")
grades <- bs_get_dataset("Grade Results")
bs_join_enrollments_grades(enrollments, grades)
}
}
}
