% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{bs_join_enrollments_orgunits}
\alias{bs_join_enrollments_orgunits}
\title{Join enrollments with org units}
\usage{
bs_join_enrollments_orgunits(enrollments, org_units)
}
\arguments{
\item{enrollments}{A tibble from the User Enrollments dataset.}

\item{org_units}{A tibble from the Org Units dataset.}
}
\value{
A joined tibble.
}
\description{
Left joins an enrollments tibble with an org units tibble on \code{org_unit_id}.
}
\examples{
\donttest{
if (bs_has_token()) {
enrollments <- bs_get_dataset("User Enrollments")
org_units <- bs_get_dataset("Org Units")
bs_join_enrollments_orgunits(enrollments, org_units)
}
}
}
