% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{bs_join_enrollments_roles}
\alias{bs_join_enrollments_roles}
\title{Join enrollments with role details}
\usage{
bs_join_enrollments_roles(enrollments, role_details)
}
\arguments{
\item{enrollments}{A tibble from the User Enrollments dataset.}

\item{role_details}{A tibble from the Role Details dataset.}
}
\value{
A joined tibble.
}
\description{
Left joins an enrollments tibble with a role details tibble on \code{role_id}.
}
\examples{
\donttest{
if (bs_has_token()) {
enrollments <- bs_get_dataset("User Enrollments")
roles <- bs_get_dataset("Role Details")
bs_join_enrollments_roles(enrollments, roles)
}
}
}
