% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{bs_join_grades_objects}
\alias{bs_join_grades_objects}
\title{Join grade results with grade objects}
\usage{
bs_join_grades_objects(grade_results, grade_objects)
}
\arguments{
\item{grade_results}{A tibble from the Grade Results dataset.}

\item{grade_objects}{A tibble from the Grade Objects dataset.}
}
\value{
A joined tibble.
}
\description{
Left joins a grade results tibble with a grade objects tibble on
\code{grade_object_id} and \code{org_unit_id}.
}
\examples{
\donttest{
if (bs_has_token()) {
grades <- bs_get_dataset("Grade Results")
objects <- bs_get_dataset("Grade Objects")
bs_join_grades_objects(grades, objects)
}
}
}
