% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{bs_join_users_enrollments}
\alias{bs_join_users_enrollments}
\title{Join users with enrollments}
\usage{
bs_join_users_enrollments(users, enrollments)
}
\arguments{
\item{users}{A tibble from the Users dataset.}

\item{enrollments}{A tibble from the User Enrollments dataset.}
}
\value{
A joined tibble.
}
\description{
Left joins a users tibble with an enrollments tibble on \code{user_id}.
}
\examples{
\dontrun{
users <- bs_get_dataset("Users")
enrollments <- bs_get_dataset("User Enrollments")
bs_join_users_enrollments(users, enrollments)
}
}
