% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ads.R
\name{bs_list_ads}
\alias{bs_list_ads}
\title{List available Advanced Data Sets}
\usage{
bs_list_ads()
}
\value{
A tibble with columns: \code{dataset_id}, \code{name}, \code{description},
\code{category}.
}
\description{
Retrieves all available ADS datasets from the Brightspace instance.
}
\examples{
\donttest{
if (bs_has_token()) {
ads <- bs_list_ads()
ads
}
}
}
