% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ads.R
\name{bs_list_ads_jobs}
\alias{bs_list_ads_jobs}
\title{List all submitted ADS export jobs}
\usage{
bs_list_ads_jobs()
}
\value{
A tibble of all submitted export jobs with columns:
\code{export_job_id}, \code{name}, \code{dataset_id}, \code{status}, \code{status_text},
\code{submit_date}.
}
\description{
List all submitted ADS export jobs
}
\examples{
\donttest{
if (bs_has_token()) {
bs_list_ads_jobs()
}
}
}
