% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ads_schemas.R
\name{bs_list_ads_schemas}
\alias{bs_list_ads_schemas}
\title{List all registered ADS dataset schemas}
\usage{
bs_list_ads_schemas()
}
\value{
A character vector of registered dataset names (snake_case).
}
\description{
List all registered ADS dataset schemas
}
\examples{
bs_list_ads_schemas()
}
