% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{bs_list_schemas}
\alias{bs_list_schemas}
\title{List all registered dataset schemas}
\usage{
bs_list_schemas()
}
\value{
A character vector of registered dataset names (snake_case).
}
\description{
List all registered dataset schemas
}
\examples{
bs_list_schemas()
}
