% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ads.R
\name{bs_org_id}
\alias{bs_org_id}
\title{Get the root organisation ID}
\usage{
bs_org_id()
}
\value{
Character string of the root org unit ID.
}
\description{
Calls \verb{/d2l/api/lp/(version)/organization/info} and returns the org
identifier.
}
\examples{
\dontrun{
bs_org_id()
}
}
