% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{bs_parse_bool}
\alias{bs_parse_bool}
\title{Parse boolean values}
\usage{
bs_parse_bool(x)
}
\arguments{
\item{x}{Character vector.}
}
\value{
Logical vector.
}
\description{
Handles Brightspace conventions: "True"/"False", "0"/"1".
}
\keyword{internal}
