% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{bs_pkce_challenge}
\alias{bs_pkce_challenge}
\title{Generate a PKCE S256 code challenge}
\usage{
bs_pkce_challenge(verifier)
}
\arguments{
\item{verifier}{The code verifier string.}
}
\value{
Base64url-encoded SHA-256 hash.
}
\description{
Generate a PKCE S256 code challenge
}
\keyword{internal}
