% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{bs_post}
\alias{bs_post}
\title{Perform a POST request with a JSON body and return parsed JSON}
\usage{
bs_post(path, body)
}
\arguments{
\item{path}{API path.}

\item{body}{A list to be serialized as JSON in the request body.}
}
\value{
Parsed JSON response as a list.
}
\description{
Perform a POST request with a JSON body and return parsed JSON
}
\keyword{internal}
