% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{bs_random_string}
\alias{bs_random_string}
\title{Generate a random string}
\usage{
bs_random_string(n = 32)
}
\arguments{
\item{n}{Number of characters.}
}
\value{
Character string.
}
\description{
Generate a random string
}
\keyword{internal}
