% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{bs_read_auth_response}
\alias{bs_read_auth_response}
\title{Read the auth redirect URL from the user}
\usage{
bs_read_auth_response()
}
\value{
Character string with the pasted URL.
}
\description{
Uses the best available input method: RStudio dialog if available,
otherwise \code{readline()}.
}
\keyword{internal}
