% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{bs_refresh_token}
\alias{bs_refresh_token}
\title{Refresh an OAuth2 token}
\usage{
bs_refresh_token(token, client_id, client_secret, scope)
}
\arguments{
\item{token}{A token list with a \code{refresh_token} field.}

\item{client_id}{OAuth2 client ID.}

\item{client_secret}{OAuth2 client secret.}

\item{scope}{OAuth2 scope.}
}
\value{
A new token list.
}
\description{
Refresh an OAuth2 token
}
\keyword{internal}
