% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{bs_retry_after}
\alias{bs_retry_after}
\title{Extract retry-after delay from response}
\usage{
bs_retry_after(resp)
}
\arguments{
\item{resp}{An httr2 response.}
}
\value{
Numeric seconds to wait.
}
\description{
Extract retry-after delay from response
}
\keyword{internal}
