% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bs_set_timezone}
\alias{bs_set_timezone}
\title{Set the timezone for Brightspace analytics}
\usage{
bs_set_timezone(tz)
}
\arguments{
\item{tz}{A valid timezone string from \code{\link[=OlsonNames]{OlsonNames()}}.}
}
\value{
Invisibly returns the timezone string.
}
\description{
All analytics functions use this timezone for converting date columns.
}
\examples{
bs_set_timezone("Pacific/Auckland")
bs_set_timezone("America/New_York")
}
