% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{bs_token_cache_path}
\alias{bs_token_cache_path}
\title{Get the token cache path}
\usage{
bs_token_cache_path(client_id)
}
\arguments{
\item{client_id}{OAuth2 client ID (used to namespace cache files).}
}
\value{
File path string.
}
\description{
Get the token cache path
}
\keyword{internal}
