% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{bs_token_needs_refresh}
\alias{bs_token_needs_refresh}
\title{Check if a token needs refreshing}
\usage{
bs_token_needs_refresh(token)
}
\arguments{
\item{token}{A token list.}
}
\value{
Logical.
}
\description{
Check if a token needs refreshing
}
\keyword{internal}
