% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bs_unzip}
\alias{bs_unzip}
\title{Extract a ZIP file to a temporary directory}
\usage{
bs_unzip(zip_path)
}
\arguments{
\item{zip_path}{Path to the ZIP file.}
}
\value{
Path to the temporary directory containing extracted files.
}
\description{
Extract a ZIP file to a temporary directory
}
\keyword{internal}
