% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ads.R
\name{format_iso8601}
\alias{format_iso8601}
\title{Format a date as ISO 8601 UTC string}
\usage{
format_iso8601(x)
}
\arguments{
\item{x}{Date, POSIXct, or character date.}
}
\value{
Character string in ISO 8601 UTC format.
}
\description{
Format a date as ISO 8601 UTC string
}
\keyword{internal}
