% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_snake_case}
\alias{to_snake_case}
\title{Convert PascalCase or mixed-case names to snake_case}
\usage{
to_snake_case(x)
}
\arguments{
\item{x}{Character vector of names.}
}
\value{
Character vector of snake_case names.
}
\description{
Convert PascalCase or mixed-case names to snake_case
}
\keyword{internal}
